fix(client): Verify we have valid entities for both doors before continuing
GetEntityHeading is very generous and doesn't error out when a nil value is passed into it, instead returning 0.0 gracefully.

My hypothesis is that there is a nasty race condition when getting the entities for double doors on line 53 where in double[1].entity gets loaded and stored, but due to latency, poor mlo construction, or something else double[2].entity has yet loaded and remains nil. The condition is still met and thus double[2].entity is no longer looked for. When attempting to toggle the lock of the door from that point forward the door that is considered double[2] is never considered closed / locked since it is nil and the heading is constantly being returned as 0.0.

I was able to verify this potentiality by forcing the second entity to not be populated and remain nil and successfully reproduced the described issue from ticket #22 .