Revert "refactor(client): remove heading check"
This reverts commit c7ef1714da199bfadf6e7d88f2e277445a92593a.

More people complain about doors not swinging (which is "intentional"
aka the game's fault) than the still yet-to-be-reproduced issue with
doors being "unlocked". If I'm going to get people being annoying
regardless let's go with the preferred behaviour.

Maybe somebody will ACTUALLY find a solution to the problem or bother
investigating when it happens to them and report solid information.